--- src/clients/client-desktop/TestingSystem.Client.Desktop/BL/BusinessServices/Windows/EndPassingTest/IEndPassingTestWindowManagement.cs
@@ -1,11 +1,0 @@
-﻿using TestingSystem.Common.BL.BusinessObjects.NonEntities;
-
-namespace TestingSystem.Client.Desktop.BL.BusinessServices.Windows.EndPassingTest
-{
-    public interface IEndPassingTestWindowManagement
-    {
-        PassingTestResultBusinessObject PassingTestResult { get; set; }
-
-        void OpenWindow();
-    }
-}
