+++ src/clients/client-desktop/TestingSystem.Client.Desktop/TestingSystem.Client.Desktop.csproj
@@ -62,25 +62,19 @@
       <DependentUpon>App.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="BL\BusinessServices\PassingTest\IPassingTestService.cs" />
-    <Compile Include="BL\BusinessServices\PassingTest\PassingTestService.cs" />
-    <Compile Include="BL\BusinessServices\PassingTest\TestQuestionsOverException.cs" />
-    <Compile Include="BL\BusinessServices\Windows\BaseWindowManagementService.cs" />
-    <Compile Include="BL\BusinessServices\Windows\EndPassingTest\EndPassingTestWindowManagement.cs" />
-    <Compile Include="BL\BusinessServices\Windows\EndPassingTest\IEndPassingTestWindowManagement.cs" />
-    <Compile Include="BL\BusinessServices\Windows\PassingTest\IPassingTestWindowManagementService.cs" />
-    <Compile Include="BL\BusinessServices\Windows\PassingTest\PassingTestWindowManagementService.cs" />
-    <Compile Include="BL\BusinessServices\Windows\SuggestedRole\ISuggestedRoleWindowManagementService.cs" />
-    <Compile Include="BL\BusinessServices\Windows\SuggestedRole\SuggestedRoleWindowManagementService.cs" />
-    <Compile Include="BL\BusinessServices\Windows\TestDetails\ITestDetailsWindowManagementService.cs" />
-    <Compile Include="BL\BusinessServices\Windows\TestDetails\TestDetailsWindowManagementService.cs" />
+    <Compile Include="BL\WindowManagement\BaseWindowManagementService.cs" />
+    <Compile Include="BL\WindowManagement\PassingTest\IPassingTestWindowManagementService.cs" />
+    <Compile Include="BL\WindowManagement\PassingTest\PassingTestWindowManagementService.cs" />
+    <Compile Include="BL\WindowManagement\SuggestedRole\ISuggestedRoleWindowManagementService.cs" />
+    <Compile Include="BL\WindowManagement\SuggestedRole\SuggestedRoleWindowManagementService.cs" />
+    <Compile Include="BL\WindowManagement\TestDetails\ITestDetailsWindowManagementService.cs" />
+    <Compile Include="BL\WindowManagement\TestDetails\TestDetailsWindowManagementService.cs" />
     <Compile Include="BL\Container\ContainerConfig.cs" />
     <Compile Include="BL\Container\ContainerModule.cs" />
     <Compile Include="BL\ViewModels\Authentication\AuthenticationLoginViewModel.cs" />
     <Compile Include="BL\ViewModels\Student\StudentDashboardViewModel.cs" />
     <Compile Include="BL\ViewModels\Student\StudentPassTestViewModel.cs" />
     <Compile Include="BL\ViewModels\Student\StudentTestDetailsViewModel.cs" />
-    <Compile Include="BL\ViewModels\Student\StudentTestPassEndViewModel.cs" />
     <Compile Include="UI\Windows\Authentication\LoginWindow.xaml.cs">
       <DependentUpon>LoginWindow.xaml</DependentUpon>
     </Compile>
